+++ app/Http/Controllers/Backend/OutletController.php
@@ -139,6 +139,11 @@ class OutletController extends Controller
      */
     public function destroy(Outlet $outlet)
     {
+        if($outlet->has('complains')) {
+            $complains = $outlet->complains->pluck('id');
+            return redirect()->route('outlet.index')->with('failure', "This outlet has following complains associated with it. Please disassociate them before deleting. $complains");
+        }
+
         try {
             $outlet->delete();
         } catch (\Exception $e) {
